refactor(server): extract mcp_registry from app/mod.rs
`build_mcp_registry()` and the supporting `ToolCategory` enum were
embedded in the 1,436-line `app/mod.rs`. They're a single
construction-time concern called once during `SkillService::new` —
isolating them keeps the dispatcher focused on lifecycle wiring.

Net change in mod.rs: -80 LOC (the function body and enum impl).
Net change overall: +18 LOC (the new module file's docstring).

Note: the larger `pub fn run()` dispatch (~590 LOC) and
`impl SkillService { ... }` (~550 LOC) blocks remain in mod.rs.
Those are higher-risk splits that deserve their own focused PRs.

480 server-lib tests pass.
